model 2/requirements.txt
opencv-python
mediapipe
deepface
pillow
pyttsx3
